// src/app/services/notifications/models/user-preference.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserPreference {
  commentMentions?: boolean;
  createdAt?: string;
  deadlineReminder?: boolean;
  emailNotifications?: boolean;
  id?: number;
  projectInvitation?: boolean;
  pushNotifications?: boolean;
  taskAssigned?: boolean;
  taskUpdated?: boolean;
  updatedAt?: string;
  userId?: number;
  websocketNotifications?: boolean;
}